+++ src/types/mod.rs
@@ -1,9 +1,11 @@
 mod calculated_population_value;
 mod player;
 mod seasonal_amount;
+mod settlement;
 mod ship;
 
 pub use calculated_population_value::CalculatedPopulationValue;
 pub use player::Player;
 pub use seasonal_amount::SeasonalAmount;
+pub use settlement::{Settlement, SettlementType};
 pub use ship::{Ship, ShipSize};
